spellblade/requirements.txt
django
djangorestframework
django-cors-headers
django-environ
djangorestframework-simplejwt
psycopg2
